Improve playlist deletion and validation logic
Added confirmation and error handling to playlist deletion in MainController. Prevented deletion and editing of playlists with id 0 in Logic. Updated PlaylistDAO to remove playlists and their songs in the correct order. Fixed method signatures and improved validation in PlaylistEditController.

Removing playlist works

+++ src/main/java/dk/easv/cs5/mytunes/bll/Logic.java
@@ -120,12 +120,19 @@ public class Logic implements ILogic {
 
     @Override
     public void deletePlaylist(Playlist playlist) {
+        if(playlist.getId() == 0){
+            return;
+        }
         playlistDAO.remove(playlist.getId());
 
     }
 
     @Override
     public void editPlaylist(Playlist playlist) {
+        if(playlist.getId() == 0){
+            return;
+        }
+
         playlistDAO.edit(playlist);
 
     }
